review cleanup
